Добавил приложение core; унаследовал модели каналов и груп от BaseWebsoketGroup
