+++ chat/users_channels/models.py
@@ -1,3 +1,12 @@
 from django.db import models
+from sorl.thumbnail import get_thumbnail
 
-# Create your models here.
+from django.utils.html import mark_safe
+
+from core.models import BaseWebsocketGroup
+
+
+class UsersChannel(BaseWebsocketGroup):
+    class Meta:
+        verbose_name = 'Канал'
+        verbose_name_plural = 'Каналы'
